+++ lib/libft/ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: rpambhar <rpambhar@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/06 14:50:48 by rpambhar          #+#    #+#             */
-/*   Updated: 2023/10/12 16:08:01 by rpambhar         ###   ########.fr       */
+/*   Updated: 2024/04/04 18:17:39 by rpambhar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_strlen(const char *s)
+int	ft_strlen(const char *s)
 {
-	size_t	len;
+	int	len;
 
 	len = 0;
 	while (s[len])
